requirements.txt
pywebview==5.4; sys_platform != 'linux'
pywebview[qt]==5.4; sys_platform == 'linux'
setuptools==70.3.0
pyinstaller==6.12.0
requests==2.32.3
